+++ backend/db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_08_23_080000) do
+ActiveRecord::Schema[8.1].define(version: 2026_08_23_090000) do
   create_table "acknowledgements", force: :cascade do |t|
     t.datetime "at", null: false
     t.datetime "created_at", null: false
@@ -231,6 +231,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_08_23_080000) do
     t.string "call_control_id"
     t.date "call_day"
     t.string "call_leg_id"
+    t.string "call_tz"
     t.datetime "completed_at"
     t.datetime "created_at", null: false
     t.integer "daily_sequence"